feat: navbar blocks (#245)
