+++ apps/website/content/blocks/nav-bar.mdx
@@ -4,70 +4,104 @@ description: 사용자의 액션에 대한 피드백 혹은 서비스에 대한 
 previewImageUrl: /images/nav-bar-preview.svg
 ---
 
-### Navbar with logo
+### Left Aligned Navbar
 
-<Demo name="default-icon-button">
-```tsx
-import { IconButton } from '@vapor-ui/core';
-import { HeartIcon } from '@vapor-ui/icons';
-
-export default function DefaultIconButton() {
-    return (
-        <IconButton>
-            <HeartIcon />
-        </IconButton>
-    );
+<Demo name="navbars/block01">
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/block01.tsx",
+    "codeblock": true
+}
+```
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/block.css",
+    "codeblock": true
 }
 ```
 </Demo>
 
-### Navbar with logo
+### Center Aligned Navbar
 
-<Demo name="default-icon-button">
-```tsx
-import { IconButton } from '@vapor-ui/core';
-import { HeartIcon } from '@vapor-ui/icons';
-
-export default function DefaultIconButton() {
-    return (
-        <IconButton>
-            <HeartIcon />
-        </IconButton>
-    );
+<Demo name="navbars/block02">
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/block02.tsx",
+    "codeblock": true
+}
+```
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/block.css",
+    "codeblock": true
 }
 ```
 </Demo>
 
-### Navbar with logo
+### Right Aligned Navbar
 
-<Demo name="default-icon-button">
-```tsx
-import { IconButton } from '@vapor-ui/core';
-import { HeartIcon } from '@vapor-ui/icons';
-
-export default function DefaultIconButton() {
-    return (
-        <IconButton>
-            <HeartIcon />
-        </IconButton>
-    );
+<Demo name="navbars/block03">
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/block03.tsx",
+    "codeblock": true
+}
+```
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/block.css",
+    "codeblock": true
 }
 ```
 </Demo>
 
-### Navbar with logo
+### Navbar with Login Button
 
-<Demo name="default-icon-button">
-```tsx
-import { IconButton } from '@vapor-ui/core';
-import { HeartIcon } from '@vapor-ui/icons';
+<Demo name="navbars/block04">
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/block04.tsx",
+    "codeblock": true
+}
+```
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/block.css",
+    "codeblock": true
+}
+```
+</Demo>
 
-export default function DefaultIconButton() {
-    return (
-        <IconButton>
-            <HeartIcon />
-        </IconButton>
-    );
+### Logo Centered Navbar
+
+<Demo name="navbars/block05">
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/block05.tsx",
+    "codeblock": true
+}
+```
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/logo.css",
+    "codeblock": true
+}
+```
+</Demo>
+
+### Page Navigation Navbar
+
+<Demo name="navbars/block06">
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/block06.tsx",
+    "codeblock": true
+}
+```
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/logo.css",
+    "codeblock": true
 }
 ```
 </Demo>
