+++ apps/website/src/components/demo/examples/navbars/logo.css
@@ -1,0 +1,3 @@
+.logo {
+    flex-shrink: 0;
+}
